requirements.txt
about-time
adagio
aiohappyeyeballs
aiohttp
aiosignal
alive-progress
alpaca-trade-api
appdirs
asgiref
asttokens
attrs
backtrader
blinker
bokeh
category-encoders
certifi
cffi
charset-normalizer
click
cloudpickle
comm
contourpy
coreforecast
cycler
Cython
darts
dash
dash-core-components
dash-html-components
dash-table
deap
decorator
deprecation
Django
et-xmlfile
executing
fastjsonschema
filelock
Flask
fonttools
frozenlist
fs
fsspec
fugue
grapheme
holidays
idna
imageio
imbalanced-learn
importlib_metadata
ipython
ipywidgets
itsdangerous
jedi
Jinja2
joblib
jsonpickle
jsonschema
jsonschema-specifications
jupyter_core
jupyterlab_widgets
kaleido
kiwisolver
lazy_loader
lightgbm
lightning-utilities
lime
llvmlite
MarkupSafe
matplotlib
matplotlib-inline
mpmath
msgpack
multidict
nbformat
nest-asyncio
networkx
nfoursid
numba
numpy
openpyxl
orjson
packaging
pandas
parso
patsy
pexpect
pillow
platformdirs
plotly
plotly-resampler
plyer
pmdarima
prompt_toolkit
protobuf
psutil
psycopg2
ptyprocess
pure_eval
pyaml
pyarrow
pycaret
pycparser
Pygments
pyod
pyparsing
python-dateutil
python-dotenv
pytorch-lightning
pytz
PyYAML
referencing
requests
retrying
rpds-py
schemdraw
scikit-base
scikit-image
scikit-learn
scikit-optimize
scikit-plot
scipy
setuptools
shap
six
sktime
slicer
sqlparse
stack-data
statsforecast
statsmodels
stopit
sympy
tbats
tenacity
tensorboardX
threadpoolctl
tifffile
torch
torchmetrics
tornado
TPOT
tqdm
traitlets
triad
tsdownsample
typing_extensions
tzdata
update-checker
urllib3
utilsforecast
vaderSentiment
vulkan
watchdog
wcwidth
websocket-client
websockets
Werkzeug
widgetsnbextension
wurlitzer
xarray
xgboost
xxhash
xyzservices
yarl
yellowbrick
zipp